refactor: use redisClusterAsyncCommandArgvToNode API
Refactored the code to use redisClusterAsyncCommandArgvToNode API which was added after hiredis-cluster 0.11.0. Furthermore, to directly use the redisClusterAsyncCommandArgv function which takes argc, argv, and argvlen as arguments. This change simplifies the code. As a result, revised the error handling mechanism by switching from "goto" to "return" statements for code clarity and maintainability.